Time: 158 ms (73.97%) | Memory: 36.6 MB (33.54%) - LeetSync

+++ 4188-divisible-game/divisible-game.cpp
@@ -41,6 +41,6 @@ public:
         }
         ll res = ((bst%mod) + mod)%mod;
         res = (res * bk)%mod;
-        return (int)res;
+        return res;
     }
 };
